SphereCast melee attacks
Changed the SphereCast on the melee attack to a SphereCastNonAlloc to allow for multiple enemies to be hit in one punch.

Stopped the player from turning during a melee attack.

Added to HealthAndDamage the ability to not be able to be killed and setting a min health to stop at (for the training dummy)